requirements.txt
beautifulsoup4>=4.12.0
supabase
python-dotenv
discord-webhook
requests
scrapeops-python-requests
python-dateutil
ratelimit
backoff
